Unify QgsMeshUtils into a single utility class
This was previously a namespace split over multiple files
