+++ python/PyQt6/core/auto_generated/mesh/qgsmeshlayerinterpolator.sip.in
@@ -12,46 +12,9 @@
 
 
 
-%ModuleHeaderCode
-#include "qgsmeshlayerinterpolator.h"
-%End
 
 
 
-
-namespace QgsMeshUtils
-{
-
-  QgsRasterBlock *exportRasterBlock(
-    const QgsMeshLayer &layer,
-    const QgsMeshDatasetIndex &datasetIndex,
-    const QgsCoordinateReferenceSystem &destinationCrs,
-    const QgsCoordinateTransformContext &transformContext,
-    double mapUnitsPerPixel,
-    const QgsRectangle &extent,
-    QgsRasterBlockFeedback *feedback = 0
-  ) /Factory/;
-%Docstring
-Exports mesh layer's dataset values as raster block
-
-The function always fetches native mesh and dataset data
-from data provider and calculates triangular mesh
-
-:param layer: mesh layer
-:param datasetIndex: index from layer defining group and dataset (time) to export
-:param destinationCrs: destination/map CRS. Used to create triangular mesh from native mesh
-:param transformContext: Transform context to transform layer CRS to destination CRS
-:param mapUnitsPerPixel: map units per pixel for block
-:param extent: extent of block in destination CRS
-:param feedback: optional raster feedback object for cancellation/preview
-
-:return: raster block with Float.64 values. ``None`` on error
-
-.. versionadded:: 3.6
-%End
-
-
-};
 
 /************************************************************************
  * This file has been generated automatically from                      *
