+++ python/PyQt6/core/auto_generated/mesh/qgsmeshutils.sip.in
@@ -1,0 +1,80 @@
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/core/mesh/qgsmeshutils.h                                         *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.py again   *
+ ************************************************************************/
+
+
+
+
+
+typedef QgsPoint QgsMeshVertex;
+
+typedef QVector<int> QgsMeshFace;
+
+typedef QPair<int, int> QgsMeshEdge;
+
+
+class QgsMeshUtils
+{
+%Docstring(signature="appended")
+Provides utility functions for working with mesh data.
+
+.. versionadded:: 3.0
+%End
+
+%TypeHeaderCode
+#include "qgsmeshutils.h"
+%End
+  public:
+
+    static QgsRasterBlock *exportRasterBlock(
+      const QgsMeshLayer &layer,
+      const QgsMeshDatasetIndex &datasetIndex,
+      const QgsCoordinateReferenceSystem &destinationCrs,
+      const QgsCoordinateTransformContext &transformContext,
+      double mapUnitsPerPixel,
+      const QgsRectangle &extent,
+      QgsRasterBlockFeedback *feedback = 0
+    ) /Factory/;
+%Docstring
+Exports mesh layer's dataset values as raster block
+
+The function always fetches native mesh and dataset data
+from data provider and calculates triangular mesh
+
+:param layer: mesh layer
+:param datasetIndex: index from layer defining group and dataset (time) to export
+:param destinationCrs: destination/map CRS. Used to create triangular mesh from native mesh
+:param transformContext: Transform context to transform layer CRS to destination CRS
+:param mapUnitsPerPixel: map units per pixel for block
+:param extent: extent of block in destination CRS
+:param feedback: optional raster feedback object for cancellation/preview
+
+:return: raster block with Float.64 values. ``None`` on error
+
+.. versionadded:: 3.6
+%End
+
+
+
+
+
+
+
+
+
+
+
+
+};
+
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/core/mesh/qgsmeshutils.h                                         *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.py again   *
+ ************************************************************************/
