+++ python/core/auto_additions/qgsmeshutils.py
@@ -1,0 +1,5 @@
+# The following has been generated automatically from src/core/mesh/qgsmeshutils.h
+try:
+    QgsMeshUtils.__group__ = ['mesh']
+except (NameError, AttributeError):
+    pass
