+++ src/core/CMakeLists.txt
@@ -854,6 +854,7 @@ set(QGIS_CORE_SRCS
   mesh/qgsmeshforcebypolylines.cpp
   mesh/qgsmeshlayerlabeling.cpp
   mesh/qgsmeshlayerlabelprovider.cpp
+  mesh/qgsmeshutils.cpp
 
   pointcloud/qgspointcloudattribute.cpp
   pointcloud/qgspointcloudattributebyramprenderer.cpp
@@ -1696,6 +1697,7 @@ set(QGIS_CORE_HDRS
   mesh/qgsmeshforcebypolylines.h
   mesh/qgsmeshlayerlabeling.h
   mesh/qgsmeshlayerlabelprovider.h
+  mesh/qgsmeshutils.h
 
   painting/qgsgeometrypaintdevice.h
   painting/qgsmaskpaintdevice.h
